Reorganize SaveCollection code to allow different Savers
Reorganize SaveCollection code to allow different Savers
